voice_forms/requirements.txt
Flask==1.1.1
textblob==0.15.3
gunicorn==20.0.4
language_tool_python==2.3.1
dnspython==2.0.0
nltk==3.4.5
